pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "rx-radar"
version = "0.2.0"
description = ""
authors = ["Rahul Chodvadiya <rahulchodvadiua@gmail.com>"]
readme = "README.md"

packages = [
    { include = "server" }
]

[tool.poetry.scripts]
start = "server.app:main"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.105.0"
uvicorn = "^0.24.0"
llama-index-core = "0.12.0"
llama-index-llms-openai = "^0.3.2"
python-dotenv = "^1.0.1"
colorama = "^0.4.6"
llama-index-llms-anthropic = "^0.5.0"
llama-index-agent-openai = "^0.4.0"
llama-index-embeddings-openai = "0.3.1"
llama-index-utils-workflow = "0.3.0"
llama-index-vector-stores-faiss = "0.3.0"
llama-index-readers-web = "0.3.0"
llama-index-readers-file = "0.4.0"
faiss-gpu-cu12 = "1.9.0.0"
python-multipart = "0.0.18"